Move the command-line ingestion runner code to a new script
